+++ lib/src/dart_formatter.dart
@@ -15,11 +15,13 @@ import 'package:analyzer/src/dart/scanner/scanner.dart';
 import 'package:analyzer/src/string_source.dart';
 import 'package:pub_semver/pub_semver.dart';
 
+import 'constants.dart';
 import 'exceptions.dart';
+import 'front_end/ast_node_visitor.dart';
+import 'short/source_visitor.dart';
+import 'short/style_fix.dart';
 import 'source_code.dart';
-import 'source_visitor.dart';
 import 'string_compare.dart' as string_compare;
-import 'style_fix.dart';
 
 /// Dart source code formatter.
 class DartFormatter {
@@ -38,6 +40,11 @@ class DartFormatter {
 
   final Set<StyleFix> fixes;
 
+  /// Flags to enable experimental language features.
+  ///
+  /// See dart.dev/go/experiments for details.
+  final List<String> experimentFlags;
+
   /// Creates a new formatter for Dart code.
   ///
   /// If [lineEnding] is given, that will be used for any newlines in the
@@ -49,17 +56,22 @@ class DartFormatter {
   ///
   /// While formatting, also applies any of the given [fixes].
   DartFormatter(
-      {this.lineEnding, int? pageWidth, int? indent, Iterable<StyleFix>? fixes})
+      {this.lineEnding,
+      int? pageWidth,
+      int? indent,
+      Iterable<StyleFix>? fixes,
+      List<String>? experimentFlags})
       : pageWidth = pageWidth ?? 80,
         indent = indent ?? 0,
-        fixes = {...?fixes};
+        fixes = {...?fixes},
+        experimentFlags = [...?experimentFlags];
 
   /// Formats the given [source] string containing an entire Dart compilation
   /// unit.
   ///
   /// If [uri] is given, it is a [String] or [Uri] used to identify the file
   /// being formatted in error messages.
-  String format(String source, {uri}) {
+  String format(String source, {Object? uri}) {
     if (uri == null) {
       // Do nothing.
     } else if (uri is Uri) {
@@ -70,7 +82,8 @@ class DartFormatter {
       throw ArgumentError('uri must be `null`, a Uri, or a String.');
     }
 
-    return formatSource(SourceCode(source, uri: uri, isCompilationUnit: true))
+    return formatSource(
+            SourceCode(source, uri: uri as String?, isCompilationUnit: true))
         .text;
   }
 
@@ -165,8 +178,15 @@ class DartFormatter {
 
     // Format it.
     var lineInfo = parseResult.lineInfo;
-    var visitor = SourceVisitor(this, lineInfo, unitSourceCode);
-    var output = visitor.run(node);
+
+    SourceCode output;
+    if (experimentFlags.contains(tallStyleExperimentFlag)) {
+      var visitor = AstNodeVisitor(this, lineInfo, unitSourceCode);
+      output = visitor.run(node);
+    } else {
+      var visitor = SourceVisitor(this, lineInfo, unitSourceCode);
+      output = visitor.run(node);
+    }
 
     // Sanity check that only whitespace was changed if that's all we expect.
     if (fixes.isEmpty &&
@@ -200,19 +220,14 @@ class DartFormatter {
   // happens to parse without error, then we use that result instead.
   ParseStringResult _parse(String source, String? uri,
       {required bool patterns}) {
-    // Enable all features that are enabled by default in the current analyzer
-    // version.
+    var version = patterns ? Version(3, 3, 0) : Version(2, 19, 0);
+
+    // Don't pass the formatter's own experiment flag to the parser.
+    var experiments = experimentFlags.toList();
+    experiments.remove(tallStyleExperimentFlag);
+
     var featureSet = FeatureSet.fromEnableFlags2(
-      sdkLanguageVersion: Version(2, 19, 0),
-      flags: [
-        'inline-class',
-        'class-modifiers',
-        if (patterns) 'patterns',
-        'records',
-        'sealed-class',
-        'unnamed-libraries',
-      ],
-    );
+        sdkLanguageVersion: version, flags: experiments);
 
     return parseString(
       content: source,