+++ test/fix_test.dart
@@ -3,7 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
-library dart_style.test.fix_test;
+library;
 
 import 'package:dart_style/dart_style.dart';
 import 'package:test/test.dart';
@@ -11,12 +11,16 @@ import 'package:test/test.dart';
 import 'utils.dart';
 
 void main() async {
-  await testFile(
-      'fixes/named_default_separator.unit', [StyleFix.namedDefaultSeparator]);
-  await testFile('fixes/doc_comments.stmt', [StyleFix.docComments]);
-  await testFile('fixes/function_typedefs.unit', [StyleFix.functionTypedefs]);
-  await testFile('fixes/optional_const.unit', [StyleFix.optionalConst]);
-  await testFile('fixes/optional_new.stmt', [StyleFix.optionalNew]);
-  await testFile('fixes/single_cascade_statements.stmt',
-      [StyleFix.singleCascadeStatements]);
+  await testFile('short/fixes/named_default_separator.unit',
+      fixes: [StyleFix.namedDefaultSeparator]);
+  await testFile('short/fixes/doc_comments.stmt',
+      fixes: [StyleFix.docComments]);
+  await testFile('short/fixes/function_typedefs.unit',
+      fixes: [StyleFix.functionTypedefs]);
+  await testFile('short/fixes/optional_const.unit',
+      fixes: [StyleFix.optionalConst]);
+  await testFile('short/fixes/optional_new.stmt',
+      fixes: [StyleFix.optionalNew]);
+  await testFile('short/fixes/single_cascade_statements.stmt',
+      fixes: [StyleFix.singleCascadeStatements]);
 }