+++ test/utils.dart
@@ -5,6 +5,8 @@
 import 'dart:io';
 
 import 'package:dart_style/dart_style.dart';
+import 'package:dart_style/src/constants.dart';
+import 'package:dart_style/src/testing/benchmark.dart';
 import 'package:dart_style/src/testing/test_file.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
@@ -127,25 +129,71 @@ Future<TestProcess> runCommandOnDir([List<String>? args]) {
   return runCommand(['.', ...?args]);
 }
 
-/// Run tests defined in "*.unit" and "*.stmt" files inside directory [name].
-Future<void> testDirectory(String name, [Iterable<StyleFix>? fixes]) async {
-  for (var test in await TestFile.listDirectory(name)) {
+/// Run tests defined in "*.unit" and "*.stmt" files inside directory [path].
+Future<void> testDirectory(String path, {Iterable<StyleFix>? fixes}) async {
+  for (var test in await TestFile.listDirectory(path)) {
     _testFile(test, fixes);
   }
 }
 
-Future<void> testFile(String path, [Iterable<StyleFix>? fixes]) async {
+Future<void> testFile(String path, {Iterable<StyleFix>? fixes}) async {
   _testFile(await TestFile.read(path), fixes);
 }
 
+/// Format all of the benchmarks and ensure they produce their expected outputs.
+Future<void> testBenchmarks({required bool useTallStyle}) async {
+  var benchmarks = Benchmark.findAll(await findPackageDirectory());
+
+  group('Benchmarks', () {
+    for (var benchmark in benchmarks) {
+      test(benchmark.name, () {
+        var formatter = DartFormatter(
+            pageWidth: benchmark.pageWidth,
+            experimentFlags: useTallStyle
+                ? const ['inline-class', 'macros', tallStyleExperimentFlag]
+                : const ['inline-class', 'macros']);
+
+        var actual = formatter.formatSource(SourceCode(benchmark.input));
+
+        // The test files always put a newline at the end of the expectation.
+        // Statements from the formatter (correctly) don't have that, so add
+        // one to line up with the expected result.
+        var actualText = actual.text;
+        if (!benchmark.isCompilationUnit) actualText += '\n';
+
+        var expected =
+            useTallStyle ? benchmark.tallOutput : benchmark.shortOutput;
+
+        // Fail with an explicit message because it's easier to read than
+        // the matcher output.
+        if (actualText != expected) {
+          fail('Formatting did not match expectation. Expected:\n'
+              '$expected\nActual:\n$actualText');
+        }
+      });
+    }
+  });
+}
+
 void _testFile(TestFile testFile, Iterable<StyleFix>? baseFixes) {
+  var useTallStyle = testFile.path.startsWith('tall/');
+
   group(testFile.path, () {
     for (var formatTest in testFile.tests) {
       test(formatTest.label, () {
+        var fixes = [...?baseFixes, ...formatTest.fixes];
+
+        if (useTallStyle && fixes.isNotEmpty) {
+          fail('Test error: Tall style does not support applying fixes.');
+        }
+
         var formatter = DartFormatter(
             pageWidth: testFile.pageWidth,
             indent: formatTest.leadingIndent,
-            fixes: [...?baseFixes, ...formatTest.fixes]);
+            fixes: fixes,
+            experimentFlags: useTallStyle
+                ? const ['inline-class', 'macros', tallStyleExperimentFlag]
+                : const ['inline-class', 'macros']);
 
         var actual = formatter.formatSource(formatTest.input);
 
@@ -160,6 +208,11 @@ void _testFile(TestFile testFile, Iterable<StyleFix>? baseFixes) {
         if (actualText != formatTest.output.text) {
           fail('Formatting did not match expectation. Expected:\n'
               '${formatTest.output.text}\nActual:\n$actualText');
+        } else if (actual.selectionStart != formatTest.output.selectionStart ||
+            actual.selectionLength != formatTest.output.selectionLength) {
+          fail('Selection did not match expectation. Expected:\n'
+              '${formatTest.output.textWithSelectionMarkers}\n'
+              'Actual:\n${actual.textWithSelectionMarkers}');
         }
 
         expect(actual.selectionStart, equals(formatTest.output.selectionStart));